Move menu button to own file to prevent circular imports
